Bring back ARM image building using ASF self-hosted ARM runners
We can now use self-hosted ARM runners to run ARM builds instead
of spinning our own instances. This PR brings back those builds to
use the ASF runners.

This still does not solve the problem of building multi-platform
images because with buildkit we need to have two interconnected
machines with two different platforms to build such an image and
this is not easy/possible with GitHub Actions. We might need to
eventually manually generate manifest and put together such
images later, but for now we build those images manually.

Related: #40925